+++ CHANGELOG.md
@@ -351,3 +351,4 @@
 - 2024-03-29: Init icp, sender, contacts, users, +2
 - 2024-03-30: Rate README, follow up, mutate, sender, +3
 - 2024-04-01: Docs summaries, site config, mod, send, +3
+- 2024-04-02: Init mod, site config, mod, summaries, +3